Refactors settings into dedicated modules
Moves settings-related code into separate modules for better organization and maintainability.

This change improves the plugin's structure by isolating settings definitions and the settings tab logic.
