+++ src/settings/index.ts
@@ -1,0 +1,4 @@
+export type { FormattedPasterPluginSettings } from './settings';
+export type { SettingsManager } from './settingsTab';
+export { DEFAULT_SETTINGS } from './settings';
+export { FormattedPasterPluginSettingTab } from './settingsTab';
